Adding workflow to uglify and build the docs.
This commit uglifies the ontology without the need of ROS, converts it
to a proper owl file, and uploads the artifacts.
This commit also adds automatic building for the specification docs.
It updates the README to contain badges for both jobs.
Last, it adds an optional step which pushes the compiled ontologies onto
the gh-pages branch if a secrets.ACCESS_TOKEN is set and the build
happened on the master branch.